combine api.js and index.js
